Aegis_AI/requirements.txt
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
scikit-learn
numpy
pandas
